Move histogram to internal/tools/
