+++ internal/tools/histogram_test.go
@@ -1,7 +1,7 @@
 // Copyright 2025 SAP SE
 // SPDX-License-Identifier: Apache-2.0
 
-package plugins
+package tools
 
 import (
 	"reflect"
